Refactor regex patterns in transformFile function to remove unused regex4
